Move RingtonePlayer and RingtonePlayerManager to new package
